+++ src/Core/Controller/Panel/DashboardController.php
@@ -24,6 +24,7 @@ use App\Core\Service\SettingService;
 use EasyCorp\Bundle\EasyAdminBundle\Config\Assets;
 use EasyCorp\Bundle\EasyAdminBundle\Config\Dashboard;
 use EasyCorp\Bundle\EasyAdminBundle\Config\MenuItem;
+use EasyCorp\Bundle\EasyAdminBundle\Config\Option\ColorScheme;
 use EasyCorp\Bundle\EasyAdminBundle\Config\UserMenu;
 use EasyCorp\Bundle\EasyAdminBundle\Controller\AbstractDashboardController;
 use Symfony\Component\HttpFoundation\Response;
@@ -67,7 +68,9 @@ class DashboardController extends AbstractDashboardController
         }
         $logo = sprintf('<img src="%s" alt="%s" style="max-width: 90%%;">', $logoUrl, $title);
         return Dashboard::new()
-            ->setTitle($logo);
+            ->setTitle($logo)
+            ->setDefaultColorScheme(ColorScheme::LIGHT)
+            ;
     }
 
     public function configureMenuItems(): iterable
@@ -118,6 +121,11 @@ class DashboardController extends AbstractDashboardController
     public function configureUserMenu(UserInterface $user): UserMenu
     {
         $userMenu = parent::configureUserMenu($user);
+        $menuItems = $userMenu->getAsDto()->getItems();
+
+        $logoutAction = end($menuItems);
+        $logoutAction->getAsDto()->setIcon('fa-sign-out-alt');
+
         $userMenu->addMenuItems([
             MenuItem::linkToCrud(
                 $this->translator->trans('pteroca.dashboard.account_settings'),
@@ -125,6 +133,7 @@ class DashboardController extends AbstractDashboardController
                 UserAccount::class
             ),
         ]);
+
         return $userMenu;
     }