+++ src/Core/Controller/Panel/Setting/AbstractSettingCrudController.php
@@ -68,7 +68,8 @@ abstract class AbstractSettingCrudController extends AbstractPanelController
                         return sprintf("%s<br><small>(%s)</small>", $value, $hint);
                     }
                     return $value;
-                }),
+                })
+                ->setHelp($this->getHelpText($this->currentEntity?->getName())),
             ChoiceField::new('type', $this->translator->trans('pteroca.crud.setting.type'))
                 ->setChoices(SettingTypeEnum::getValues())
                 ->onlyWhenCreating(),
@@ -194,5 +195,17 @@ abstract class AbstractSettingCrudController extends AbstractPanelController
             return $this->settingRepository->find($id);
         }
         return null;
+    }
+
+    private function getHelpText(?string $name): string
+    {
+        if (empty ($name)) {
+            return '';
+        }
+
+        $hintIndex = "pteroca.crud.setting.hints.$name";
+        $hint = $this->translator->trans($hintIndex);
+
+        return $hint !== $hintIndex ? $hint : '';
     }
 }