+++ src/Core/Service/Pterodactyl/PterodactylAccountService.php
@@ -3,6 +3,7 @@
 namespace App\Core\Service\Pterodactyl;
 
 use App\Core\Entity\User;
+use Exception;
 use Timdesm\PterodactylPhpApi\Resources\User as PterodactylUser;
 
 class PterodactylAccountService
@@ -16,13 +17,41 @@ class PterodactylAccountService
 
     public function createPterodactylAccount(User $user, string $plainPassword): PterodactylUser
     {
-        return $this->pterodactylService->getApi()->users->create([
-            'email' => $user->getEmail(),
-            'username' => $this->usernameService->generateUsername($user->getEmail()),
-            'first_name' => $user->getName(),
-            'last_name' => $user->getSurname(),
-            'password' => $plainPassword,
-        ]);
+       try {
+           return $this->pterodactylService->getApi()->users->create([
+               'email' => $user->getEmail(),
+               'username' => $this->usernameService->generateUsername($user->getEmail()),
+               'first_name' => $user->getName(),
+               'last_name' => $user->getSurname(),
+               'password' => $plainPassword,
+           ]);
+       } catch (Exception $e) {
+           $deepErrors = $e->errors['errors'] ?? [];
+           $errors = array_map(fn($error) => $error['detail'], $deepErrors);
+           $error = sprintf(
+               'Error during creating Pterodactyl account: %s (%s)',
+               $e->getMessage(),
+               implode(', ', $errors)
+           );
+
+           throw new Exception($error);
+       }
+    }
+
+    public function updatePterodactylAccountPassword(User $user, string $plainPassword): PterodactylUser
+    {
+        try {
+            $currentPterodactylUser = $this->pterodactylService->getApi()->users->get($user->getPterodactylUserId());
+            return $this->pterodactylService->getApi()->users->update($user->getPterodactylUserId(), [
+                'email' => $currentPterodactylUser->get('email'),
+                'username' => $currentPterodactylUser->get('username'),
+                'first_name' => $currentPterodactylUser->get('first_name'),
+                'last_name' => $currentPterodactylUser->get('last_name'),
+                'password' => $plainPassword,
+            ]);
+        } catch (Exception $e) {
+            throw new Exception(sprintf('Error while updating user password: %s', $e->getMessage()));
+        }
     }
 
     public function deletePterodactylAccount(User $user): void