+++ src/Core/Service/StoreService.php
@@ -69,6 +69,11 @@ class StoreService
             $product->setImagePath($imagePath . $product->getImagePath());
         }
 
+        if (!empty($product->getBannerPath())) {
+            $bannerPath = $this->productsBasePath . '/';
+            $product->setBannerPath($bannerPath . $product->getBannerPath());
+        }
+
         return $product;
     }
 
@@ -117,9 +122,13 @@ class StoreService
             throw new NotFoundHttpException($this->translator->trans('pteroca.store.egg_not_found'));
         }
 
+        $this->validateUserBalance($user, $product);
+    }
 
+    public function validateUserBalance(User $user, Product $product): void
+    {
         if (($product->getPrice() / 100) > $user->getBalance()) {
-            throw new NotFoundHttpException($this->translator->trans('pteroca.store.not_enough_funds'));
+            throw new \Exception($this->translator->trans('pteroca.store.not_enough_funds'));
         }
     }
 }