+++ src/Core/Twig/AppExtension.php
@@ -2,8 +2,10 @@
 
 namespace App\Core\Twig;
 
+use App\Core\Entity\User;
 use App\Core\Enum\SettingEnum;
 use App\Core\Service\SettingService;
+use Symfony\Component\Routing\RouterInterface;
 use Twig\Extension\AbstractExtension;
 use Twig\TwigFunction;
 
@@ -12,6 +14,7 @@ class AppExtension extends AbstractExtension
     public function __construct(
         private readonly string $currentVersion,
         private readonly SettingService $settingService,
+        private readonly RouterInterface $router,
     ) {}
 
     public function getFunctions(): array
@@ -28,6 +31,7 @@ class AppExtension extends AbstractExtension
             new TwigFunction('get_favicon', [$this, 'getFavicon']),
             new TwigFunction('use_pterodactyl_panel_as_client_panel', [$this, 'usePterodactylPanelAsClientPanel']),
             new TwigFunction('get_pterodactyl_panel_url', [$this, 'getPterodactylPanelUrl']),
+            new TwigFunction('is_pterodactyl_sso_enabled', [$this, 'isPterodactylSSOEnabled']),
         ];
     }
 
@@ -86,9 +90,27 @@ class AppExtension extends AbstractExtension
         return (bool)$this->settingService->getSetting(SettingEnum::PTERODACTYL_PANEL_USE_AS_CLIENT_PANEL->value);
     }
 
-    public function getPterodactylPanelUrl(): string
+    public function isPterodactylSSOEnabled(): bool
     {
-        return $this->settingService->getSetting(SettingEnum::PTERODACTYL_PANEL_URL->value);
+        return (bool)$this->settingService->getSetting(SettingEnum::PTERODACTYL_SSO_ENABLED->value);
+    }
+
+    public function getPterodactylPanelUrl(string $path = ''): string
+    {
+        $isPterodactylSSOEnabled = $this->settingService->getSetting(SettingEnum::PTERODACTYL_SSO_ENABLED->value);
+        $pterodactylPanelUrl = $isPterodactylSSOEnabled
+            ? $this->router->generate('sso_redirect')
+            : $this->settingService->getSetting(SettingEnum::PTERODACTYL_PANEL_URL->value);
+
+        if (!empty($path)) {
+            if (!$isPterodactylSSOEnabled) {
+                $pterodactylPanelUrl = rtrim($pterodactylPanelUrl, '/') . '/' . ltrim($path, '/');
+            } else {
+                $pterodactylPanelUrl .= '?redirect_path=' . $path;
+            }
+        }
+
+        return $pterodactylPanelUrl;
     }
 
     public function getCaptchaSiteKey(): ?string